Cambio de tipo documento beneficiario

See merge request supergiros/proyecto-de-autogestion!36

+++ src/main/java/com/supergiros/portalautogestion/domain/TransaccionesNomina.java
@@ -26,7 +26,7 @@ public class TransaccionesNomina implements Serializable {
     private String tipoDocumentoBenef;
 
     @Column(name = "numero_documento_benef")
-    private Integer numeroDocumentoBenef;
+    private Long numeroDocumentoBenef;
 
     @Column(name = "nombre_uno")
     private String nombreUno;
@@ -140,16 +140,16 @@ public class TransaccionesNomina implements Serializable {
         this.tipoDocumentoBenef = tipoDocumentoBenef;
     }
 
-    public Integer getNumeroDocumentoBenef() {
+    public Long getNumeroDocumentoBenef() {
         return this.numeroDocumentoBenef;
     }
 
-    public TransaccionesNomina numeroDocumentoBenef(Integer numeroDocumentoBenef) {
+    public TransaccionesNomina numeroDocumentoBenef(Long numeroDocumentoBenef) {
         this.setNumeroDocumentoBenef(numeroDocumentoBenef);
         return this;
     }
 
-    public void setNumeroDocumentoBenef(Integer numeroDocumentoBenef) {
+    public void setNumeroDocumentoBenef(Long numeroDocumentoBenef) {
         this.numeroDocumentoBenef = numeroDocumentoBenef;
     }
